+++ R/step-first.R
@@ -61,6 +61,9 @@
 #'   filter(mpg < mean(mpg)) %>%
 #'   summarise(hp = mean(hp))
 lazy_dt <- function(x, name = NULL, immutable = TRUE, key_by = NULL) {
+  # in case `x` has an `as.data.table()` method but not a `group_vars()` method 
+  groups <- tryCatch(group_vars(x), error = function(e) character())
+
   if (!is.data.table(x)) {
     if (!immutable) {
       abort("`immutable` must be `TRUE` when `x` is not already a data table.")
@@ -80,7 +83,7 @@ lazy_dt <- function(x, name = NULL, immutable = TRUE, key_by = NULL) {
     data.table::setkeyv(x, key_vars)
   }
 
-  step_first(x, name = name, immutable = immutable, env = caller_env())
+  step_first(x, name = name, groups = groups, immutable = immutable, env = caller_env())
 }
 
 #' @export
@@ -88,7 +91,8 @@ dim.dtplyr_step_first <- function(x) {
   dim(x$parent)
 }
 
-step_first <- function(parent, name = NULL, immutable = TRUE, env = caller_env()) {
+step_first <- function(parent, name = NULL, groups = character(),
+                       immutable = TRUE, env = caller_env()) {
   stopifnot(is.data.table(parent))
 
   if (is.null(name)) {
@@ -97,7 +101,7 @@ step_first <- function(parent, name = NULL, immutable = TRUE, env = caller_env()
 
   new_step(parent,
     vars = names(parent),
-    groups = character(),
+    groups = groups,
     locals = list(),
     implicit_copy = !immutable,
     needs_copy = FALSE,